move InventoryKind out of valence_protocol
