+++ crates/valence_new/src/inventory.rs
@@ -6,7 +6,7 @@ use tracing::warn;
 use valence_protocol::packets::s2c::play::{
     CloseContainerS2c, OpenScreen, SetContainerContentEncode, SetContainerSlotEncode,
 };
-use valence_protocol::{InventoryKind, ItemStack, Text, VarInt, WindowType};
+use valence_protocol::{ItemStack, Text, VarInt, WindowType};
 
 use crate::client::event::CloseContainer;
 use crate::client::Client;
@@ -158,6 +158,134 @@ pub(crate) fn update_player_inventories(mut query: Query<(&mut Inventory, &mut C
                 slot_idx: -1,
                 slot_data: cursor_item.as_ref(),
             });
+        }
+    }
+}
+
+#[derive(Copy, Clone, PartialEq, Eq, Debug)]
+pub enum InventoryKind {
+    Generic9x1,
+    Generic9x2,
+    Generic9x3,
+    Generic9x4,
+    Generic9x5,
+    Generic9x6,
+    Generic3x3,
+    Anvil,
+    Beacon,
+    BlastFurnace,
+    BrewingStand,
+    Crafting,
+    Enchantment,
+    Furnace,
+    Grindstone,
+    Hopper,
+    Lectern,
+    Loom,
+    Merchant,
+    ShulkerBox,
+    Smithing,
+    Smoker,
+    Cartography,
+    Stonecutter,
+    Player,
+}
+
+impl InventoryKind {
+    /// The number of slots in this inventory. When the inventory is shown to
+    /// clients, this number does not include the player's main inventory slots.
+    pub const fn slot_count(self) -> usize {
+        match self {
+            InventoryKind::Generic9x1 => 9,
+            InventoryKind::Generic9x2 => 9 * 2,
+            InventoryKind::Generic9x3 => 9 * 3,
+            InventoryKind::Generic9x4 => 9 * 4,
+            InventoryKind::Generic9x5 => 9 * 5,
+            InventoryKind::Generic9x6 => 9 * 6,
+            InventoryKind::Generic3x3 => 3 * 3,
+            InventoryKind::Anvil => 4,
+            InventoryKind::Beacon => 1,
+            InventoryKind::BlastFurnace => 3,
+            InventoryKind::BrewingStand => 5,
+            InventoryKind::Crafting => 10,
+            InventoryKind::Enchantment => 2,
+            InventoryKind::Furnace => 3,
+            InventoryKind::Grindstone => 3,
+            InventoryKind::Hopper => 5,
+            InventoryKind::Lectern => 1,
+            InventoryKind::Loom => 4,
+            InventoryKind::Merchant => 3,
+            InventoryKind::ShulkerBox => 27,
+            InventoryKind::Smithing => 3,
+            InventoryKind::Smoker => 3,
+            InventoryKind::Cartography => 3,
+            InventoryKind::Stonecutter => 2,
+            InventoryKind::Player => 45,
+        }
+    }
+}
+
+impl From<InventoryKind> for WindowType {
+    fn from(value: InventoryKind) -> Self {
+        match value {
+            InventoryKind::Generic9x1 => WindowType::Generic9x1,
+            InventoryKind::Generic9x2 => WindowType::Generic9x2,
+            InventoryKind::Generic9x3 => WindowType::Generic9x3,
+            InventoryKind::Generic9x4 => WindowType::Generic9x4,
+            InventoryKind::Generic9x5 => WindowType::Generic9x5,
+            InventoryKind::Generic9x6 => WindowType::Generic9x6,
+            InventoryKind::Generic3x3 => WindowType::Generic3x3,
+            InventoryKind::Anvil => WindowType::Anvil,
+            InventoryKind::Beacon => WindowType::Beacon,
+            InventoryKind::BlastFurnace => WindowType::BlastFurnace,
+            InventoryKind::BrewingStand => WindowType::BrewingStand,
+            InventoryKind::Crafting => WindowType::Crafting,
+            InventoryKind::Enchantment => WindowType::Enchantment,
+            InventoryKind::Furnace => WindowType::Furnace,
+            InventoryKind::Grindstone => WindowType::Grindstone,
+            InventoryKind::Hopper => WindowType::Hopper,
+            InventoryKind::Lectern => WindowType::Lectern,
+            InventoryKind::Loom => WindowType::Loom,
+            InventoryKind::Merchant => WindowType::Merchant,
+            InventoryKind::ShulkerBox => WindowType::ShulkerBox,
+            InventoryKind::Smithing => WindowType::Smithing,
+            InventoryKind::Smoker => WindowType::Smoker,
+            InventoryKind::Cartography => WindowType::Cartography,
+            InventoryKind::Stonecutter => WindowType::Stonecutter,
+            // arbitrarily chosen, because a player inventory technically does not have a window
+            // type
+            InventoryKind::Player => WindowType::Generic9x4,
+        }
+    }
+}
+
+impl From<WindowType> for InventoryKind {
+    fn from(value: WindowType) -> Self {
+        match value {
+            WindowType::Generic9x1 => InventoryKind::Generic9x1,
+            WindowType::Generic9x2 => InventoryKind::Generic9x2,
+            WindowType::Generic9x3 => InventoryKind::Generic9x3,
+            WindowType::Generic9x4 => InventoryKind::Generic9x4,
+            WindowType::Generic9x5 => InventoryKind::Generic9x5,
+            WindowType::Generic9x6 => InventoryKind::Generic9x6,
+            WindowType::Generic3x3 => InventoryKind::Generic3x3,
+            WindowType::Anvil => InventoryKind::Anvil,
+            WindowType::Beacon => InventoryKind::Beacon,
+            WindowType::BlastFurnace => InventoryKind::BlastFurnace,
+            WindowType::BrewingStand => InventoryKind::BrewingStand,
+            WindowType::Crafting => InventoryKind::Crafting,
+            WindowType::Enchantment => InventoryKind::Enchantment,
+            WindowType::Furnace => InventoryKind::Furnace,
+            WindowType::Grindstone => InventoryKind::Grindstone,
+            WindowType::Hopper => InventoryKind::Hopper,
+            WindowType::Lectern => InventoryKind::Lectern,
+            WindowType::Loom => InventoryKind::Loom,
+            WindowType::Merchant => InventoryKind::Merchant,
+            WindowType::ShulkerBox => InventoryKind::ShulkerBox,
+            WindowType::Smithing => InventoryKind::Smithing,
+            WindowType::Smoker => InventoryKind::Smoker,
+            WindowType::Cartography => InventoryKind::Cartography,
+            WindowType::Stonecutter => InventoryKind::Stonecutter,
         }
     }
 }
