+++ crates/valence_new/src/server.rs
@@ -15,7 +15,7 @@ use tokio::runtime::{Handle, Runtime};
 use tokio::sync::{OwnedSemaphorePermit, Semaphore};
 use uuid::Uuid;
 use valence_nbt::{compound, Compound, List};
-use valence_protocol::{ident, InventoryKind, Username};
+use valence_protocol::{ident, Username};
 
 use crate::biome::{Biome, BiomeId};
 use crate::client::event::{dispatch_client_events, register_client_events};
@@ -26,7 +26,7 @@ use crate::entity::{deinit_despawned_entities, init_entities, McEntityManager};
 use crate::instance::{update_instances_post_client, update_instances_pre_client, Instance};
 use crate::inventory::{
     handle_close_container, update_client_on_close_inventory, update_client_on_open_inventory,
-    update_open_inventories, update_player_inventories, Inventory,
+    update_open_inventories, update_player_inventories, Inventory, InventoryKind,
 };
 use crate::player_textures::SignedPlayerTextures;
 use crate::server::connect::do_accept_loop;
